refactor(runtime): split orchestrator into focused submodules
Extract three submodules from the 1,540-line orchestrator/mod.rs:

- prompt.rs: system prompt composition and skill XML rendering
- dispatch.rs: tool dispatch, confirmation gate, result finalization
- turn_control.rs: end_turn evaluation and FinalAnswer handling

mod.rs retains the Orchestrator struct, constructor, public turn API,
and the two turn-loop implementations. No behavioral changes.
